test: Run the E2E test locally
