+++ test/request.ts
@@ -1,0 +1,3 @@
+// For now, you'll need to do something like this to get a correctly-typed
+// `Request` to pass to `worker.fetch()`.
+export const IncomingRequest = Request<unknown, IncomingRequestCfProperties>;
